Fix deleting LB managed by controller
#28 introduced two bugs related to deleting an ALB managed by the
controller. First bug was attempting to detach the target group from the
ASG twice. Second bug was not initializing the http listener struct when
getting an ALB, this meant the http listener was never deleted from the
ALB and thus the ALB deletion failed because of the active listener
resource.

This fixes both of those issues.