Add support for legacy KubernetesCluster tag
This adds support for the legacy KubernetesCluster tag as a fallback
when detecting clusterID from the host instance.

This is to support kops which still use the old tag for the ec2
instances.

The new tag kubernetes.io/cluster/<cluster-id> will still be used and
expected by the ingress controller for the security group and the stack
that is created.