Predictably select subnets
This changes the logic of selecting subnets for the ALB to match the logic used
by the kube-controller-manager when selecting subnets for ELBs used for
services of type LoadBalancer.

This solves several problems:

1. The controller will no longer try to create an ALB with multiple subnets of
the same AZ. (Fixes #108)

2. If only some subnets are tagged with the elb role the controller will favour
those with the tag but not limit the selection to only those with a tag.

3. Favour using internal subnets for ALBs with the scheme `internal`.

Fix #108