Initial work for zalando-incubator/kube-ingress-aws-controller#38
Implements a new flag -ingress-class-filter which if specified is
interpreted as a comma seperated list of values which will be used
as a filter to match kubernetes.io/ingress.class annotations. If
the annotation is not set, or does match, the ingress will be hidden
from the ingress controller and not acted upon.

As of this commit the work passes a 'make test' but requires
further testing.