add optional support for cloudwatch alarms
This adds support for cloudwatch alarm configuration for each load
balancer created by the controller. This is an opt-in feature. To use
it, the `--cloudwatch-alarms-config-map` CLI flag has to be set and
should be pointed to a ConfigMap that contains the alarm configuration.
If the ConfigMap does not exist or contains malformed data, the
controller will just ignore it. Check out to `cloudwatch.md` for a full
documentation of the feature.

Signed-off-by: Martin Ohmann <martin.ohmann@bonial.com>