Restore CUSTOM_FILTERS function which was broken in 0.8.4.
This fixes #266, which describes the issue being addressed here. In short, CUSTOM_FILTERS is only applied to DescribeInstances, so ensure we limit the ASGs considered to those seen in the described instances.

It was originally considered to try to apply tag filtering to the ASG lookups as well, but in the end, simply restored the prior logic with a minimal patch.

Signed-off-by: Joe Hohertz <joe@viafoura.com>