ALB Listener rule to deny internal domain requests
This commit addresses #402. It adds new the following new configuration
flags:
  - internal-domains
  - deny-internal-domains
  - deny-internal-domains-response
  - deny-internal-domains-response-content-type
  - deny-internal-domains-response-status-code

The main flag is the `deny-internal-domains`, a boolean that enables and
disables the feature. The other flags allow to set the internal domains
that will be protected as well as the response that will be returned.
Defaults are also set: an HTTP `text/plain` `401 - Unauthorized`
response for the `*.ingress.cluster.local` domain.

The ALB cloudformation template will add a
[`AWS::ElasticLoadBalancingV2::ListenerRule`][0] resource, configured
with the [condition][1] values from the `internal-domains` and the
[action `fixedresponseconfig`][2] with the respective response
`deny-internal-domains-response` flags.

[0]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-elasticloadbalancingv2-listenerrule.html
[1]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticloadbalancingv2-listenerrule-hostheaderconfig.html
[2]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticloadbalancingv2-listenerrule-action.html#cfn-elasticloadbalancingv2-listenerrule-action-fixedresponseconfig

Signed-off-by: Jonathan Juares Beber <jonathanbeber@gmail.com>