Add the ability to use an HTTPS target
This adds a boolean flag called `target-https` which will set the
protocol used on a target group to HTTPS.

*Motivation*

AWS VPC networking does prevent external targets from sniffing the wire,
but if the target is a proxy listening on the hostNetwork, any other pod
deployed to that node that is also using the hostNetwork can read the traffic
unencrypted. In cases where end-to-end encryption is expected or
desirable, this allows for that option.

Signed-off-by: Jeremy Huntwork <jhuntwork@lightcubesolutions.com>