CNI: validate initial pod list
There is a corner case where the initial listing of the pods happens during an ingress pod start phase. A pod without an assigned IP address might be included then, causing an ALB TG update error.

We are validating the pod statuses in the informer event handler but not yet at the initial listing. Adding the same validations there as well.

Signed-off-by: Samuel Lang <gh@lang-sam.de>