Add goldenfile tests to ingress-controller worker
What?

Adds tests in the golden file format, where given an ingress definition
in the format of a k8s yaml and compare the cloud formation template
generated by the controller, based on this k8s yaml and a mocked AWS
state, and compares it to the content of a .cf file defined in the test.

This should work like an integration test between the different adapters
and controller logic.

Why?

The introduction of these tests will bring more confidence when making
changes to the ingress controller which is a critical component of our
infrastructure. This also enable us to add new test cases if we catch
problems or anomalies increasing confidence that the problem would not
happen again.

Is there any breaking changes?

No.

-----

Related issues:

- https://github.bus.zalan.do/pg9/issues/issues/340

Signed-off-by: Lucas Thiesen <lucas.thiesen@zalando.de>