Expose custom cert pool for test purposes
What?

This commit adds a method to expose the package global variable roots.
This variable enable us to define a custom x509 certificate pool to
verify certificates against.

Why?

This change is necessary to enable the testability of the controller and
worker. To test a self signed certificate there is only two options,
given the structure of the code:

1 - Add the CA certificate to the mechine or VM running the tests and
  configuring the OS through env vars to use this CA.

2 - Verify the certificate against a customs certificate pool containing
  the CA certificate used to generate our server certificate.

Analysing these two options it was decided that the scond approach would
be better for maintainability and readability.

Is there any breaking changes?

No.

-----

Related issues:
- https://github.bus.zalan.do/pg9/issues/issues/340

Reference links:

- https://pkg.go.dev/crypto/x509
- https://www.rfc-editor.org/rfc/rfc2459.html

Signed-off-by: Lucas Thiesen <lucas.thiesen@zalando.de>