fix: update ingress status only if the load balancer is active (#748)
As per the existing behavior, once the stack update completes, the
ingress controller immediately updates the status of all ingresses and
routegroups to reference the new load balancer. This update may
sometimes happens before the new load balancer has marked its targets
(for example skipper-ingress) as healthy, leading clients being routed to a load
balancer that is not ready to serve traffic yet.

To improve this behavior we retrieve the ELBs via AWS API and build the
models including the ELB states of the corresponding ELBs. Then before
updating the ingresses/routegroups (updateIngress func) we check whether
the ELB is in active state, if not we skip updating the
ingresses/routegroups status with the ELB hostname.

---------

Signed-off-by: Thilina Madumal <thilina.madumal@zalando.de>