Enable Support for creating Dual Stack NLB (#766)
This PR Enables Support for creating Dual Stack NLB.

This is required for Enabling AWSCNI mode in EKS IPv6 clusters. For
AWSCNI mode to be enabled in an IPv6 cluster we need to have IPv6 target
groups. And IPv6 target groups can only be attached to a `dualstack`
load balancer.

There is already a flag `--ip-addr-type` to control the ALB IP address
type but for NLB we hard code it to `ipv4`, this PR just removes the
hardcoding and passes the flag value.

---------

Signed-off-by: speruri <surya.srikar.peruri@zalando.de>
Co-authored-by: speruri <surya.srikar.peruri@zalando.de>